Refining the strucuted object monolith.
